fix(SummedAreaTable): remove new conversion warnings on clang

+++ SrcLib/visu/fwRenderOgre/include/fwRenderOgre/vr/SummedAreaTable.hpp
@@ -91,7 +91,7 @@ private:
     int m_passOrientation;
 
     /// The index of the slice to which we currently render.
-    int m_sliceIndex;
+    size_t m_sliceIndex;
 
     /// The read offset based on the number of reads per fragment shader (r) and the pass index (i) : m_readOffset =
     /// r^i.
